pyproject.toml
[project]
name = "personal-ai-employee"
version = "0.1.0"
description = "Personal AI Employee - Autonomous Digital FTE"
requires-python = ">=3.12"
dependencies = [
    # AI Providers
    "anthropic>=0.40.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.35.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    # MCP Protocol
    "mcp>=1.0.0",
    # Browser Automation
    "playwright>=1.40.0",
    # Core Utilities
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pandas>=2.1.0",
    # File Watching
    "watchdog>=3.0.0",
    # Web & Scheduling
    "flask>=3.0.0",
    "schedule>=1.2.0",
    # Data Formats
    "pyyaml>=6.0.2",
    "markdown>=3.7",
    # OAuth for Twitter MCP
    "oauthlib>=3.2.0",
    "requests-oauthlib>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp", "watchers", "scripts"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
asyncio-mode = "auto"
testpaths = ["tests"]
